requirements.txt
arabic-reshaper==3.0.0
asgiref==3.11.1
asn1crypto==1.5.1
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
cssselect2==0.9.0
Django==5.2.12
django-extensions==4.1
fonttools==4.62.1
freetype-py==2.5.1
html5lib==1.1
idna==3.11
lxml==6.0.2
oscrypto==1.3.0
pillow==12.1.1
pycairo==1.29.0
pycparser==3.0
pydyf==0.12.1
pyHanko==0.34.1
pyhanko-certvalidator==0.30.1
pypdf==6.9.1
pyphen==0.17.2
python-bidi==0.6.7
PyYAML==6.0.3
reportlab==4.4.10
requests==2.32.5
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.5
svglib==1.6.0
tinycss2==1.5.1
tinyhtml5==2.1.0
tzdata==2025.3
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.3
weasyprint==68.1
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.4.1
psycopg2-binary==2.9.10
django-environ==0.12.0
channels==4.2.0
daphne==4.1.2

